docs: 파일명 변경 및 input 컴포넌트 사용법 주석 추가
