+++ src/components/Input.tsx
@@ -1,0 +1,73 @@
+import React from 'react';
+import { global } from '@/styles/global.css';
+import * as styles from './Input.css';
+
+/** input 컴포넌트: 아이콘을 포함한 input 컴포넌트입니다.
+ * Props는 선택사항이며, 필요에 따라 사용하면 됩니다.
+ * 아이콘이 있는 경우, placeholder와 겹칠 수 있으므로 inline style로 padding을 조절해주세요.
+ * 
+ * <Props 설명> 
+ * 1. icon: 아이콘 컴포넌트
+ * 2. iconPosition: 아이콘 위치
+ * 3. borderRadius: 기본이 0.4rem
+ * 4. padding: 기본이 1.6rem 1.5rem
+ * 
+ * <사용법>
+ * 1. 아이콘이 없는 경우
+ * <Input placeholder='입력' />
+ * 
+ * 2. 아이콘이 있는 경우
+ * <Input
+          style={{ paddingLeft: '1.2rem' }}
+          placeholder='내가 원하는 체험은'
+          icon={<Search />}
+          iconPosition='left'
+        />
+ * 3. 비밀번호 입력 시, 아이콘 변경
+        <Input
+          type={type}
+          value={value}
+          onChange={(e) => setValue(e.target.value)}
+          style={{ paddingRight: '1.2rem' }}
+          borderRadius='0.6rem'
+          padding='1.6rem 2rem'
+          placeholder='비밀번호를 입력해주세요'
+          icon={
+            type === 'password' ? (
+              <VisibilityOff onClick={() => setType('text')} />
+            ) : (
+              <VisibilityOn onClick={() => setType('password')} />
+            )
+          }
+          iconPosition='right'
+        />
+ */
+
+interface IconInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  icon?: React.ReactNode;
+  iconPosition?: 'left' | 'right';
+  borderRadius?: '0.4rem' | '0.6rem';
+  padding?: '1.6rem 1.5rem' | '1.6rem 2rem';
+}
+
+export default function Input(props: IconInputProps) {
+  const { icon, iconPosition, borderRadius = '0.4rem', padding = '1.6rem 1.5rem', ...others } = props;
+
+  if (!others.type) others.type = 'text';
+
+  others.style = {
+    fontSize: global.fontSize.small,
+    lineHeight: global.lineHeights.small,
+    color: global.color.gray[600],
+    fontWeight: '400',
+    ...others.style,
+  };
+
+  return (
+    <div className={styles.inputContainer} style={{ borderRadius, padding }}>
+      <div className={styles.leftIconDiv}>{icon && iconPosition === 'left' && icon}</div>
+      <input className={styles.inputField} {...others} />
+      <div className={styles.rightIconDiv}>{icon && iconPosition === 'right' && icon}</div>
+    </div>
+  );
+}
